# Compiler performance optimizations