PR message TODO

All of the changes here were authored by @aconsroe-hpe in https://github.com/Cray/chapel-private/issues/2834


Hello World compile times - built with quickstart + CHPL_LLVM=system - showing times from two runs
 * main 0432111a06 4.48 / 4.54 / 4.47 s
 * inline Symbol::qualType 4.49 / 4.51 s
 * ResolutionCandidate adjustment 4.48 / 4.51 s
 * simplify ResolveScope 4.48 / 4.54 s
 * optimize normalize 4.43 / 4.48 s

For the normalize opt here are compile times for the fft-timecomp compile
 * before: 7.46 / 7.50 / 7.53 s
 * after: 7.5 / 7.46 / 7.46 s